% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_truncated_exp.R
\name{solve_truncated_exponential}
\alias{solve_truncated_exponential}
\title{Solve for the rate parameter of a truncated exponential distribution}
\usage{
solve_truncated_exponential(
  desired_mean,
  a = 0,
  b = Inf,
  initial_guess = 1/(desired_mean - a),
  verbose = FALSE
)
}
\arguments{
\item{desired_mean}{The desired mean of the truncated exponential distribution.}

\item{a}{Lower value of the truncation. Default is \eqn{0}.}

\item{b}{Upper value of the truncation. Default is \eqn{\infty}.}

\item{initial_guess}{Initial guess for the rate parameter. Default is \eqn{1 / (desired\_mean - a)}.}

\item{verbose}{Prints output of solver if TRUE.}
}
\value{
Returns the rate parameter found.
}
\description{
Solve for the rate parameter of a truncated exponential distribution
}
\examples{
## Example use for desired mean 3 with lower truncation at 1 and upper
## truncation at 100
solve_truncated_exponential(1, 3, 100)

## Print output of optim by setting verbose=TRUE
solve_truncated_exponential(1, verbose = TRUE)
}
