% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_trucated_gamma.R
\name{solve_truncated_gamma}
\alias{solve_truncated_gamma}
\title{Solve for parameters of a truncated gamma distribution}
\usage{
solve_truncated_gamma(
  desired_mean,
  x_value,
  desired_prob,
  a = 0,
  b = Inf,
  initial_guess = c(2, 1),
  verbose = FALSE
)
}
\arguments{
\item{desired_mean}{The desired mean of the truncated gamma distribution.}

\item{x_value}{The value where the CDF is evaluated.}

\item{desired_prob}{The desired value of the CDF at \code{x_value}.}

\item{a}{Lower value of the truncation. Default is \eqn{0}.}

\item{b}{Upper value of the truncation. Default is \eqn{\infty}.}

\item{initial_guess}{Initial guess for the numerical solver, in the form of shape and rate.}

\item{verbose}{Prints output of solver if TRUE.}
}
\value{
Returns a list of the parameters found.
}
\description{
Solve for parameters of a truncated gamma distribution
}
\examples{
## Example use for desired mean 2 and P(X<=1)=0.3 with default truncation
solve_truncated_gamma(2, 1, 0.3)

## Print output of optim by setting verbose=TRUE
solve_truncated_gamma(2, 1, 0.3, verbose = TRUE)
}
